refactor: introduce cli crate with semantic help headings
Extract Cli, AnchorSubcommands, version statics, and style helper from
anchor/src/main.rs into a new anchor/cli library crate, establishing a
single source of truth for CLI type definitions.

Replace the blanket FLAG_HEADER = "Flags" with per-group heading constants
and add #[command(next_help_heading)] annotations to option group structs
so that `anchor node --help` displays semantically grouped sections.
