+++ anchor/Cargo.toml
@@ -18,8 +18,8 @@ spec-minimal = []
 async-channel = { workspace = true }
 bls = { workspace = true }
 clap = { workspace = true }
+cli = { workspace = true }
 client = { workspace = true }
-ethereum_hashing = "0.7.0"
 futures = { workspace = true }
 global_config = { workspace = true }
 keygen = { workspace = true }
@@ -31,7 +31,6 @@ tracing = { workspace = true }
 tracing-appender = { workspace = true }
 tracing-subscriber = { workspace = true }
 types = { workspace = true }
-version = { workspace = true }
 
 [dev-dependencies]
 assert_cmd = "2.0"
