+++ anchor/cli/build.rs
@@ -1,0 +1,4 @@
+fn main() {
+    // This build script exists solely to ensure that `OUT_DIR` is set at compile time,
+    // which is used by `build_profile_name()` to determine the build profile.
+}
